requirements.txt
Flask
requests
beautifulsoup4
feedgen
python-dateutil
